[PLAT-17037] Skip running UpdateConsistencyCheck for Failover
Summary: During the failover task, the source universe might be inaccessible and the task has to wait until `UpdateConsistencyCheck` subtask times out. By skipping the `UpdateConsistencyCheck` subtask, we reduce the failover task execution time by about 3 minutes.

Test Plan: Created a DR and brought down the source universe and did a failover. Made sure that the `UpdateConsistencyCheck` subtask was not run and the execution time was about 3 minutes faster.

Reviewers: amindrov, spothuraju, yshchetinin, daniel

Reviewed By: amindrov

Differential Revision: https://phorge.dev.yugabyte.com/D42537