[PLAT-17039] Ignore running TransferXClusterCerts on the source universe during failover
Summary: To further improve the failover performance, we ignore running the `createTransferXClusterCertsRemoveTasks` subtask on the source universe during failover task. It will leave the target universe root certificate on the source universe, however, it will do no harm having that extra file on the source universe. On the other hand, if the source universe is not accessible, it can reduce the execution time of the failover task by about 20 seconds.

Test Plan: Created a DR and brought down the source universe and did a failover. Made sure that the `createTransferXClusterCertsRemoveTasks` subtask was not run for the source universe and the execution time was 18 seconds faster.

Reviewers: amindrov, spothuraju, yshchetinin, daniel

Reviewed By: amindrov

Differential Revision: https://phorge.dev.yugabyte.com/D42545