[#26376] YSQL: Use `pgsql_proxy_bind_address` to identify the pg socket if available
Summary:
In certain cases `pgsql_proxy_bind_address` is set to a different value from `rpc_bind_addresses`. Most notable it is set to `0.0.0.0`.
The socket used by PG uses this the value of this flag when available. So use this value to compute the pg socket address for YSQL major upgrade.

Fixes #26376
Jira: DB-15727

Test Plan: Tested manually on YBM cluster

Reviewers: telgersma

Reviewed By: telgersma

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D42463