[PLAT-17029] XCluster bootstrap tests failing with row mismatch after restart replication
Summary:
YBA does not reset the value for the backup/restore objects relation to the xCluster config object in the subtasks as it is set and computed during parent task execution.

Since we have added retry-ability to the create xCluster task, we skip the backup/restore if the object associated with the corresponding table/namespace exists and is in completed status. Since the subtasks didn't reset the relation for the backup/restore during the remove subtasks, it leads to an issue where backup/restore is skipped.

This diff makes changes to make sure the backup/restore object relation to the xCluster object is reset at the right place at the parent level task execution.

Test Plan: - Made sure a restart xCluster task will do backup/restore even though there is a relation between the xCluster config object and backup/restore objects while the backup/restore objects are in completed status.

Reviewers: amindrov, spothuraju, yshchetinin, skhilar, daniel

Reviewed By: amindrov

Differential Revision: https://phorge.dev.yugabyte.com/D42590