[#26399] YSQL: Fix regression bug with incremental cache refresh
Summary:
Commit ac0c421a5ca7676d7ad0f73c90dd812c7eb7d7cb added a new unit test
`PgCatalogVersionTest.WaitForSharedCatalogVersionToCatchup` that demonstrated a
regression bug when `--TEST_ysql_yb_enable_invalidation_messages=true`.

The root cause is the well-known heartbeat delay problem: after a breaking DDL
that increments the breaking catalog version, it takes a heartbeat delay
interval before the new catalog version shows up in shared memory that can be
accessed by a PG backend. Although the heartbeat delay affects correctness
regardless of the value of `--TEST_ysql_yb_enable_invalidation_messages`, there
is a regression bug when multiple DDLs were executed in the following scenario:

```
session 1: (main script)             session 2: (ddl script)

SELECT * FROM foo;
\! ddl_script
                                     GRANT ALL ON SCHEMA public TO PUBLIC;
                                     REVOKE USAGE ON SCHEMA public FROM PUBLIC;
SELECT * FROM foo;
```

In session 1, assume the local catalog version is x, we first can do a
successful `SELECT` query with local catalog version x. Then in session 1 we
start a session 2 via the ysqlsh syntax `\!`. When session 2 is executed
session1 is blocked. In session 2, two DDLs are executed: `GRANT` increments the
current catalog version to x + 1 but is not a breaking DDL, `REVOKE` is a
breaking DDL that increments the breaking catalog version to x + 2. The
regression bug happens once session 2 completes and session 1 resumes and if the
new catalog version x + 1 of `GRANT` has reached shared memory of session 1 when
the second `SELECT` is executed, but the new catalog version x + 2 of `REVOKE`
has not arrived yet:

When `--TEST_ysql_yb_enable_invalidation_messages=false`, before doing the full
catalog cache refresh, we make a RPC to master to retrieve the latest catalog
version x + 2 and update session 1's local catalog version to x + 2. The second
`SELECT` can run successfully.

When `--TEST_ysql_yb_enable_invalidation_messages=true`, we only apply the
invalidation messages of x + 1, but not x + 2. After that we update session 1's
local catalog version to x + 1 only. Then when executing the second `SELECT`, it
is possible that the new x + 2 has reached the target tserver of the `SELECT`,
since (x + 1) < (x + 2), the second `SELECT` fails with
```
ERROR:  The catalog snapshot used for this transaction has been invalidated:
```

To fix the bug, I changed the code so that when doing incremental catalog cache
refresh I also make an additional RPC to master to retrieve the latest catalog
version. If it is newer than the shared catalog version, we know that the shared
catalog version is already stale and if we just move forward we are likely to
encounter the above error. To prevent that error I added a new function
`YbWaitForSharedCatalogVersionToCatchup` so that we can wait for the shared
catalog version to reach x + 2. Then we move forward with applying the
invalidation messages of x + 1 and x + 2. After that the local catalog version
will be updated to x + 2. We will not see the regression bug described above.

Note that once we have object locking and we can make the DDL wait instead of
letting PG side to wait: the DDL cannot complete (via releasing the object lock)
until x + 2 has propagated to the shared memory of all tservers. So when session
2 completes, it is guaranteed that session 1 will see x + 2 in shared memory and
therefore the regression bug will also disappear. In that case we can remove
this fix to avoid the additional master RPC call to improve performance.

In addition, when a DDL is executed and the new version returned isn't exactly x
+ 1 we cannot do the optimization to update local catalog version to x + 1. In
this case we know there is another concurrent DDL has incremented catalog
version to x + 1, if we blindly update local catalog version to x + 1 we will
miss the invalidation messages associated with x + 1. So we leave the local
catalog version at x. But since we already know x is stale since the just
completed DDL has returned `new_version` as the latest catalog version, instead
of moving forward with a stale x, I also call `YbWaitForSharedCatalogVersionToCatchup`
to wait for the `new_version` to show up in shared memory. Then we can apply the
invalidation messages up to `new_version` and update local catalog version to
`new_version` before executing the next top-level SQL statement.

The unit test `PgCatalogVersionTest.WaitForSharedCatalogVersionToCatchup` is
updated to reflect this fix.

Test Plan: ./yb_build.sh --cxx-test pg_catalog_version-test   --gtest_filter PgCatalogVersionTest.WaitForSharedCatalogVersionToCatchup -n 20 --tp 1

Reviewers: kfranz, sanketh, mihnea

Reviewed By: kfranz

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D42594