[PLAT-17057][Master] Observed UI flickering in task tabs during DB upgrade
Summary:
The refetching logic continues to trigger until the updatingTaskUUID is updated by the backend, causing the frontend to juggle.

We initially added this logic in this [[ https://phorge.dev.yugabyte.com/D42379 | diff ]] to update the banner in the new TaskDetails UI. Now, we have modified it to run only when the newTaskDetailsUI is enabled. Additionally, we maintain a state for the refetch logic.

Test Plan: Tested manually with both NewTaskDetailsuI enabled and disabled

Reviewers: lsangappa

Reviewed By: lsangappa

Differential Revision: https://phorge.dev.yugabyte.com/D42606