[BACKPORT 2024.2][#26419] YSQL: Introduce custom ParameterStatus packets for Connection Manager
Summary:
This diff introduces a custom ParameterStatus (ConnMgrParameterStatus) packet to send from the server to with Connection Manager. The purpose of this ConnMgrParameterStatus packet is to signal to Connection Manager to not send back the packet to the client, while still functioning like a ParameterStatus packet to be absorbed by Connection Manager. This in turn will allow Connection Manager to maintain protocol level correctness with respect to how PG operates, rather than bombarding the client with ParameterStatus packets which get sent specifically for Connection Manager to store the GUC state of a client.

Original commit: 882b2ac9f1067dae5911fdbd67be540a174adbfb / D42515

**Merge Conflicts**
guc.c - On master/pg15, there is an addition condition introduced when compared to 2024.2/pg11 before sending back ParameterStatus packets to the client. The introduction of the ConnMgrParameterStatus packet modified some code around this condition, leading to some merge conflicts that had to be fixed.
Jira: DB-15781

Test Plan: Jenkins: all tests, enable connection manager

Reviewers: skumar, mkumar, devansh.saxena

Reviewed By: skumar

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D42611