[PLAT-17038] Reduce the timeout for DeleteReplicationOnSource during failover
Summary:
This makes changes so the user can set the timeout for the `DeleteReplicationOnSource` subtask during failover to reduce the execution time for the failover task. When the source universe is down, this RPC call will timeout, the lower the timeout, the faster the task will complete and the user can use the dr replica as the primary universe.

If the user set the runtime `yb.xcluster.db_scoped.failover.delete_replication_on_source_timeout` to 0, it will skip that subtask for failover task.

Test Plan:
Made sure the `DeleteReplicationOnSource` subtask is skipped, when the runtime config value is set to 0. Failover task took 28 seconds to finish
With the default value of 10s for the timeout, it took 50 seconds to finish the failover task, reducing the execution time from 158 seconds while the timeout is 120 seconds which is the current default timeout for the yb-admin client.

Reviewers: amindrov, spothuraju, yshchetinin, jmak, daniel

Reviewed By: spothuraju

Differential Revision: https://phorge.dev.yugabyte.com/D42581