[#26272] CDC: Populate wal_status in pg_replication_slots
Summary:
Currently we have hardcoded `wal_status` field to lost in `pg_replication_slots` view.

With this revision we are setting the value of `wal_status` as lost, i.e., declaring the stream as expired if the CDC stream has not been consumed (via the GetChanges RPC) within the past `cdc_intent_retention_ms` duration which is a Tserver GFlag, otherwise we are setting it as `reserved`. The way we determine the stream's last_active_time is same as the way it is determined for checking whether a replication slot is active or inactive.

**Upgrade/Rollback safety:**
This diff adds a field `expired` to the message PgReplicationSlotInfoPB. New field is added to the tserver response (to pggate) proto. This is upgrade and rollback safe. No new flags are added to guard the feature.
Jira: DB-15617

Test Plan:
./yb_build.sh --java-test 'org.yb.pgsql.TestPgReplicationSlot#testActivePidAndWalStatusPopulationOnStreamRestart'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgReplicationSlot#testWalStateLost'

Reviewers: skumar, vkushwaha, sumukh.phalgaonkar, utkarsh.munjal

Reviewed By: sumukh.phalgaonkar

Subscribers: sumukh.phalgaonkar, yql

Differential Revision: https://phorge.dev.yugabyte.com/D42239