[BACKPORT 2024.2][PLAT-17071] Fix UI RBAC check for restart replication and repair dr
Summary:
In the repair DR modal and the restart repicaiton banner, the RBAC validator wrapper component
failed to provide a resource UUID. This resulted in the action being blocked despite the user
having correct permissions.

Original commit: 64b375cbb0894cee50fb661cc84dc6918691ec8d / D42635

Test Plan:
On a user with admin permissions, verify the user can restart replication or choose a new target
from the repair DR modal.
{F341711}
On a user with xCluster permissions only on a subset of universes, verify the user is unable to
choose a universe which they don't have xCluster permissions on as the new target universe.
{F341712}

Reviewers: kkannan, vbansal, hzare, lsangappa

Reviewed By: kkannan

Differential Revision: https://phorge.dev.yugabyte.com/D42855