[BACKPORT 2024.2][PLAT-16972] Minor fix on schedule code where incremental backup next time is updated before full backup next time is updated
Summary:
Original commit: 92c9dac975ef5b0435ec77a63c9e8b17f3c00870 / D42381
Next full backup time was updated but not reflected in method where next incremental backup
time is updated. This practically does not happen in code because it requires next incremental
backup time to be null.

The method we use to update the next incremental backup task time is:
```
ScheduleUtil.nextExpectedIncrementTaskTime
```
where we pass the current instance of schedule. It uses the schedule's `nextScheduleTaskTime`
but since the instance's value was not updated the getter would not know about it.
So we need to update the `nextScheduleTaskTime` first.

Test Plan: Added UT

Reviewers: asharma, vbansal

Reviewed By: asharma

Differential Revision: https://phorge.dev.yugabyte.com/D42643