[PLAT-17104] Set default scope for OIDC config
Summary:
After the SSO library upgrade, it no longer uses the default scope automatically so we set it on our end.
Also added fail safe if the runtime key is accidently set to empty string.

Test Plan:
Manual verified that login fails if scope is unset.
Verified that login works with default scope and if its set to empty string.

Reviewers: svarshney

Reviewed By: svarshney

Differential Revision: https://phorge.dev.yugabyte.com/D42684