[#26494] docdb: Remove client.h from most headers
Summary:
`client.h` takes a couple seconds to parse each time it is included and it is transitively included in many unnecessary headers. Removing this usage and replacing it with forward declarations speeds up builds about 83s.

The only non-test headers `client.h` is now used in are `client-internal.h`, `client_builder-internal.h`, `yb-admin_client.h`, and `client_master_rpc.h`. None of these files are themselves included in any header files.

This diff also adds the `-ftime-trace` flag to the top-level `CMakeLists.txt` when building with Clang. This causes Clang to output the time it took to compile each file next to the object file in the build folder as a json file. The resulting trace can be visualized with chrome://tracing.
Jira: DB-15861

Test Plan:
This was measured on an M1 laptop with 10 cores from the first compilation command to the start of initdb (not including initdb itself), skipping tests and java. The full command was:
```
./yb_build.sh --clang17 --no-ccache --skip-tests --skip-java --clean
```

This diff improves build time as measured above from 810s to 727s.

Reviewers: hsunder, #db-approvers

Reviewed By: hsunder, #db-approvers

Subscribers: slingam, svc_phabricator, yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D42673