[#26523] Packaging: Add liblz4 to extra_postgres_libs
Summary:
After we started building postgres `--with-lz4`, the YB client packaging
started to fail on macOS due to the unexpected `liblz4` dependency inside
`pg_rewind`.

This diff adds `liblz4.1.dylib` to the list of extra postgres libs so
it can copy it to the temp folder for packaging.
Jira: DB-15890

Test Plan:
```
./yb_release --package_name yugabyte-client
```

Reviewers: jharveysmith

Reviewed By: jharveysmith

Differential Revision: https://phorge.dev.yugabyte.com/D42732