[#23785] YSQL: Add a unit test for invalidation message stability
Summary:
This diff adds a new unit test that checks the stability of invalidation
messages generated by a set of common DDL statements. The common DDL statements
are chosen by consulting online PG doc and examples and looking at our existing
tests. This test can be expanded to cover more DDLs when needed.

The idea is that from release to release, we may see same DDL generates
different list of invalidation messages. If that happens we expect this unit
test to fail and then we should do a deeper inspection to see whether we need to
increment `yb_version` of one or more types of `SharedInvalidationMessage`.

The test is written by setting
`--ysql_yb_invalidation_message_expiration_secs=36000` so that we do not clear
the `pg_yb_invalidation_messages` table. It also disable auto-analyze explicitly
to prevent any unexpected ANALYZE statements from disturbing the test result. In
this way all the invalidation messages generated by the set of DDL statements
will be retained and at the end of the test, we do a simple `SELECT` of the
entire contents of `pg_yb_invalidation_messages` and then compute a fingerprint.
The test fails if the fingerprint changes.

Test Plan: ./yb_build.sh --cxx-test pg_catalog_version-test   --gtest_filter PgCatalogVersionTest.InvalMessageSampleDDls -n 10

Reviewers: kfranz, sanketh, mihnea

Reviewed By: kfranz

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D42767