[#26496] YSQL: redefine CurrentMemoryContext for upstream merge
Summary:
When comparing src/postgres with upstream PG, one of the main
differences is CurrentMemoryContext changing to
GetCurrentMemoryContext().  This is for YB-specific multithreading
support.  To preserve the behavior but reduce changes with upstream PG,
redefine CurrentMemoryContext to this function call.  The definition
changes are as follows:

- CurrentMemoryContext global variable -> YbCurrentMemoryContext
- GetCurrentMemoryContext function -> YbGetCurrentMemoryContext
- Add new CurrentMemoryContext macro defined to
  YbGetCurrentMemoryContext()

The usage changes are as follows:

- GetCurrentMemoryContext() -> CurrentMemoryContext
- CurrentMemoryContext -> YbCurrentMemoryContext

This increases differences with upstream on the setting side but reduces
differences on the getting side.  This is worth it since there are only
a handful of sets but a lot of gets.

Jira: DB-15863

Test Plan:
Close: #26496
Jenkins: compile only

Reviewers: amartsinchyk

Reviewed By: amartsinchyk

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D42729