[PLAT-17144] Put create/edit/delete CipherTrust KMS behind a runtime config on UI
Summary:
Create/edit/delete CipherTrust KMS is not possible if the runtime config
`yb.kms.allow_ciphertrust` is false.

This diff removes the CipherTrust KMS option from the list of KMS providers
in the creation form when the `yb.kms.allow_ciphertrust` runtime config flag
is false. It also disables the edit KMS and delete KMS action on
existing CipherTrust KMS configs when the flag is false.

Test Plan:
Create a CipherTrust KMS config with the runtime config set to true.
{F343100}
Set the `yb.kms.allow_ciphertrust` runtime config to false.
Verify users are unable to create CipherTrust KMS configs when the runtime config
is set to false.
{F343101}
Verify that the edit and delete actions are disabled with a tooltip indicating
why the action is disabled.
{F343099}
{F343098}
Verify that the edit and delete actions are not disabled for other KMS providers.
{F343097}

Reviewers: skurapati, asathyan, kkannan

Reviewed By: skurapati

Differential Revision: https://phorge.dev.yugabyte.com/D42769