[#26548] docdb: Add metrics to track intentdb write thread JoinBatchGroup wait time
Summary:
For each intentdb, there is one write thread responsible for writing transaction data to
the memory table and one cleanup thread that removes intentdb entries in the background.
Both threads compete to become the leader of the write batch group. Only one leader at a
time; when one thread becomes the leader, the other thread must wait. We should add a
metric to track the duration that the write thread and the remove thread spend waiting
to join the group.
Jira: DB-15916

Test Plan: manual test the metrcis is available

Reviewers: sergei, rthallam

Reviewed By: sergei

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D42772