[PLAT-17146]Overwrite check in restore targets DBs not being restored
Summary:
During a restore operation (Restore Entire Backup), we retrieve the list of keyspaces from the backup through a preflight check and from the database via the /tables API. We then determine the intersection of these lists to identify duplicates.

However, in the case of a single keyspace restore, instead of comparing only the specified keyspace, the comparison is performed across all keyspaces. This causes the form to display duplicates even if selected keyspace is absent and other tables are present.

This update introduces checks to filter out unnecessary keyspaces when performing a single keyspace restore.

Test Plan: Tested manually on both YSQL and YCQL db's. New PITR restore modal is also checked.

Reviewers: lsangappa

Reviewed By: lsangappa

Differential Revision: https://phorge.dev.yugabyte.com/D42777