[PLAT-17138] Change PG restore dump file copy to persisted location
Summary:
Observed that during restore in K8s YBA with `--skip_dump_file_delete` enabled, we were copying over
to `/opt/yugabyte` which is not a persisted location. Instead we should be copying over to
`/opt/yugabyte/yugaware/data` which is where the PVC is mounted. Fixed it in this PR.

Turned the `k8s_restore_skip_dump_file_delete` off by default. I will enable it once tested.

ALso, we should not be restoring PG after restart for Standby promotion cases since we mark the
instance as active after restore and that will be lost if we restore after restart. So we should do it during the
restore itself.

Test Plan: Verified the standby promotion is working as expected.

Reviewers: muthu

Reviewed By: muthu

Differential Revision: https://phorge.dev.yugabyte.com/D42811