[#26575] YSQL: Resolve YB_TODO for pg_stat_statements in pg_dump
Summary:
There's a work-around in `dumpACL` (pg_dump.c) to avoid dumping the ACL for pg_stat_statements_reset. Without this work-around, the upgrade fails with:
```
[m-1] 2025-03-27 10:43:16.146 EDT [97719] ERROR:  function pg_catalog.pg_stat_statements_reset() does not exist
[m-1] 2025-03-27 10:43:16.146 EDT [97719] STATEMENT:  SELECT pg_catalog.binary_upgrade_set_record_init_privs(true);
[m-1] 	REVOKE ALL ON FUNCTION "pg_catalog"."pg_stat_statements_reset"() FROM PUBLIC;
[m-1] 	SELECT pg_catalog.binary_upgrade_set_record_init_privs(false);
```

Postgres does not dump object definitions for built-in extensions, only their permissions. In YB, pg_stat_statements is a built-in
extension that's created as part of initdb. As a result, the pg_dump during binary upgrade does not contain the definition for the function `pg_stat_statements_reset`, but it contains an ACL dump for it. In the latest version of pg_stat_statements (1.10), this function's signature has changed (it has gained parameters). Since the reinitdb during upgrade creates the latest version of the extension, the REVOKE command above fails as `pg_stat_statements_reset()` no longer exists.

For now, keep the work-around and update the comment for the YB_TODO. In the future, we may want to follow the typical extension binary upgrade  path for pg_stat_statements (create an empty extension and manually create extension objects), and then this work-around can be removed (tracked in GH issue #26566).
Jira: DB-15943

Test Plan: Jenkins: compile only

Reviewers: telgersma, hsunder

Reviewed By: telgersma, hsunder

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D42796