[#25911] YSQL: Fix constraint error when moving rows between partitions
Summary:
This commit fixes an issue where attempting
to move a row across partitions of the same
partitioned table would result in a constraint
violation on the source partition.
This happened because, while updating a
row, constraint violations on the source
partition were checked before determining
whether the row would remain in it after
the update.

Changes:
1. Constraint check now happens after
determining to which partition the row
will belong after the update.
2. Added a Java Test (TestPgUpdate).

Fixes #25911

Test Plan:
./yb_build.sh --java-test org.yb.pgsql.TestPgUpdate