[#23785] YSQL: Turn on incremental catalog cache refresh by default
Summary:
To get more test exposure in master branch and make some test tasks easier. This
diff turns on incremental catalog cache refresh by default by setting the
default value of the preview gflag `--ysql-yb_enable_invalidation_messages` to
true. It is planned that before we cut the 2025.1 branch, we revert this diff if
we needed more testing time. In case we find wide impact bug that affects master
branch testing we can quickly revert this diff as well.

There is one test log needs to be updated in this diff, due to the way we handle
ANALYZE and optimize away catalog version increment for DDL that does not
generate any invalidation messages.

Test Plan: jenkins

Reviewers: kfranz, sanketh, mihnea

Reviewed By: kfranz

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D42821