[PLAT-16982] Force universe destroy is blocked by DeleteBootstrapIds failure
Summary:
This diff updates the code to not throw the exception in the `DeleteBootstrapIds` subtask if the `forceDelete` option is true.

Also, it removes the required annotation for the `bootstrapParams.tables` field because if the user set `allowBootstrap` to true, `bootstrapParams.tables` can be empty.

Test Plan: - Made sure the create xCluster API can be called with `bootstrapParams.tables` empty.

Reviewers: amindrov, spothuraju, yshchetinin, jmak, daniel

Reviewed By: jmak

Differential Revision: https://phorge.dev.yugabyte.com/D42850