[#26456] YSQL: Fix "Duplicate table" error caused by relfilenode collision
Summary:
Commit ed4b4d2633c46b36cda01971949bd3522f058104 made a fix in an attempt to
avoid "Duplicate table" error caused by OID collision. When performing a
database restore operation, certain OIDs need to remain the same. This means
that some OIDs are already reserved and cannot be reused by the OID allocator.

It turned out that the previous fix did not work always correctly if we need to
reserve `relfilenode` which are also OIDs during database restore operation.
When allocating a relfilenode, ed4b4d2633c46b36cda01971949bd3522f058104 ensures
that a newly allocated OID did not already exist in the oid column of `pg_class`
catalog table, that is not right. It should ensure that the new OID did not
already exist in the relfilenode column of `pg_class`.

This diff undo the code fix of ed4b4d2633c46b36cda01971949bd3522f058104 (but
retained the unit test), and redo the fix by checking whether a newly allocated
OID already exists in the `relfilenode` column of `pg_class` catalog table.
Because `pg_class` has no index on `relfilenode` column, a table scan is
performed to read the entire set of `relfilenode` column values and store that
in a PG hash table (HTAB). This HTAB is only built at most once during a PG
session when it is needed for the first time.

Note that when allocating for table's OID, we also need this HTAB. In PG when a
table is created for the first time, the table's OID is used both as the OID in
pg_class for this table, also as the relfilenode for this table. As a result,
the first time we need to allocate table's OID or relfilenode in a session, this
HTAB is built and retained for the entire session.

The negative performance impact is that if a session executes DDL statements,
the first DDL statement that requires OID allocation for a new table oid or
relfilenode will spend time to build the HTAB. Additional DDLs should not see
noticeable performance slowdown because HTAB provides efficient lookups.

Also note that we do not need to consider any new changes to `relfilenode`
column values in pg_class caused by on-going DDL statements because the OID
allocator ensures that any newly allocated `relfilenode` cannot cause OID
collision. Only those reserved `relfilenode` can cause OID collision.

A new unit test is added that would fail before this diff.

Test Plan:
./yb_build.sh --cxx-test pgwrapper_pg_libpq-test --gtest_filter PgLibPqTest.TableRewriteOidCollision
YB_EXTRA_TSERVER_FLAGS="--ysql_pg_conf_csv=log_statement=all" ./yb_build.sh --cxx-test pgwrapper_pg_libpq-test --gtest_filter PgLibPqTest.RelfilenodeOidCollision

Reviewers: fizaa, mihnea

Reviewed By: fizaa

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D42877