[BACKPORT 2024.2][PLAT-17156] Fix create universe page crash due to reading storageType from null
Summary:
In `StorageTypeField.tsx`, `masterFieldValue` can be null when the user opts to
placing YB-Masters on dedicated nodes in the create universe form.
Prior to this change, the page would crash when accessing `storageType`.

This change addresses the issue by adding optional chaining so we do not try to
read `storageType` if `masterFieldValue` is null.

Original Commit: ce0bd99d09e8908ddba0892e7d8b6180f0c5b4df / D42868

Test Plan:
Create GCP universe with YB-Masters on dedicated nodes.
Verify that the storageType field is populated in the form and the page does not crash.
{F344029}

Reviewers: yshchetinin, kkannan, lsangappa

Reviewed By: kkannan

Differential Revision: https://phorge.dev.yugabyte.com/D43006