[#26611] YSQL: disable PgIndexBackfillTest.GinStress on mac
Summary:
C++ test PgIndexBackfillTest.GinStress is flaky on mac.  Since this test
is sensitive to the performance of the machine, and the mac machines
that jenkins runs has performance issues, without further look, disable
the test entirely on mac.  So the remaining coverage for this test is on
linux.

In order to support doing this, add a new macro
YB_LINUX_RELEASE_ONLY_TEST.  Since it is related, rename
YB_DISABLE_TEST_EXCEPT_RELEASE to YB_RELEASE_ONLY_TEST without changing
the behavior (though the code is inverted to look closer to
YB_LINUX_RELEASE_ONLY_TEST).

Jira: DB-15987

Test Plan:
jenkins

Close: #26611

Reviewers: telgersma

Reviewed By: telgersma

Subscribers: telgersma, yql

Differential Revision: https://phorge.dev.yugabyte.com/D42869