[#26536] YSQL: Fix crash in single row check when leftop in non-Var
Summary:
yb_single_row_update_or_delete_path() incorrectly assumes that leftop of the WHERE clause is a Var. Remove this assumption using the following logic:

Single row path is applicable iff the WHERE clause specifies only the values of all the PK columns. So, if the leftop is not a Var, single row path is not possible.
Jira: DB-15899

Test Plan:
   ./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressPushdown'

Reviewers: kramanathan

Reviewed By: kramanathan

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D42809