[PLAT-17206] Increase JUnit timeout for xCluster local provier tests
Summary: We have a global timeout for the local provider tests of 900 seconds. From the recent Jenkins job, we have seen that the XCluster/DR local provider tests time out every now and then. This diff increase the time for for the XCluster/DR local provider tests to 1800 seconds to deflake the utest pipeline.

Test Plan: Run the utests

Reviewers: nsingh, amindrov, yshchetinin, daniel, nbhatia

Reviewed By: amindrov

Differential Revision: https://phorge.dev.yugabyte.com/D42882