[#26590] YSQL, TerminatedQueries: Adding query_id field to the yb_terminated_queries view.
Summary:
Adding the query_id field to the yb_terminated_queries view.

The new YbTerminatedQuery structure consists of the fields:-

1. UserID
2. DatabaseOID
3. Backend PID
4. Query ID (newly added)
5. Query Start Time
6. Query Termination Time
7. Query Text
8. Termination Reason

We get the query_id of the terminated query from the backend entry where it is a unsigned int64, but we store it as a signed int64 while adding it to the YbTerminatedQuery structure.

Therefore query_id for a query can be negative.
Jira: DB-15961

Test Plan:
./yb_build.sh --java-test 'org.yb.pgsql.TestPgYbStat'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressYbStat'
./yb_build.sh --cxx-test pgwrapper_pg_libpq-test --gtest_filter PgLibPqTest.TestLWPgBackendKillAfterLWLockAcquire

Reviewers: hbhanawat, asaha, ishan.chhangani, telgersma

Reviewed By: asaha, telgersma

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D42889