[#26605] YSQL: Add RPC to get tserver pg socket dir
Summary:
During YSQL Major upgrade test, the pg_upgrade process needs to connect to the v11 Postgres hosted by the TServer.
When the TServer is running on the same node as the Master, we use the socket dir to connect to the Postgres instance. The socket dir can be different from the reported proxy host port in certain scenarios.
This change adds a new RPC to get the socket dir directly from the TServer.

**Upgrade/Rollback safety:**
New RPC is optional. YB-Master uses the `ERROR_NO_SUCH_METHOD` to determine if the Tserver it connects to does not support this RPC, and falls back to the existing mechanism of using Master gFlag overrides.
Jira: DB-15976

Test Plan:
PgWrapperTest.GetPgSocketDir

The following tests are added, but will be disabled until the upgrade from version is bumped to include this change:
YsqlMajorUpgradeTestWithConnMgr.SimpleTableUpgrade
YsqlMajorUpgradeTestWithConnMgr.SimpleTableRollback

Reviewers: telgersma, fizaa

Reviewed By: telgersma

Subscribers: yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D42871